Generate selector entity in popUp function, components and all. Begin transfering logic of selecting particular elements from player (and items) to selector (and options).
Learned that the data passed in the 'HitOn' event is identical to the convoluted way I had borrowed from the previously cited repo. Also learned that when skipping the selector from option to option, the HitOn event does not refire unless first it recognizes that it has left the option (which it does not do from option to option.

I'm thinking I'll be able to achieve this by passing additional parameters (dynamic ones) into the checkHits method, but in the meantime, I've added a resetHitChecks method to the HitOn event, which works fine. The only reason I'm not fully satisfied with it is because my console log line begins repeating itself as the function fires again and again. I suppose it's fine for MVP/prototype, though. Moving on!.