Destroy popUp and all related entities when selection is made with ENTER. Unfreeze player entity.
Player now just starts moving again as soon as the selection is made, rather than waiting for user input. This tells me that my freeze method is not idea. I may be able to circumnavigate this issue by not using the fourway shorthand method, and instead writing the logic for KeyDown and KeyUp movement, manually. That way, with KeyUp, there should be no movement. We shall see.