Comment out freeze calls for events. May be better to...
place freeze method inside firing conditional so that the player doesn't just randomly disappear (by the trigger of the prompt method without the popup box because of the conditional). Freeze was initially put in place because hitting the selector would make the player move, which may prove to be a problem.